[task5] Two versions of service
