+++ tasks/task5/booking-service/main_v1.go
@@ -1,0 +1,30 @@
+package main
+
+import (
+  "fmt"
+  "log"
+  "net/http"
+  "os"
+)
+
+func main() {
+  enableFeatureX := os.Getenv("ENABLE_FEATURE_X") == "true"
+
+  http.HandleFunc("/ping", func(w http.ResponseWriter, r *http.Request) {
+    fmt.Fprintf(w, "pong")
+  })
+
+  http.HandleFunc("/ready", func(w http.ResponseWriter, r *http.Request) {
+    fmt.Fprintf(w, "ready")
+  })
+
+  // if ENABLE_FEATURE_X=true, expose /feature
+  if enableFeatureX {
+    http.HandleFunc("/feature", func(w http.ResponseWriter, r *http.Request) {
+      fmt.Fprintf(w, "Feature X is enabled!")
+    })
+  }
+
+  log.Println("Server v1 running on :8080")
+  log.Fatal(http.ListenAndServe(":8080", nil))
+}
